pyproject.toml
[project]
name = "igbo-archives-agent-hq"
version = "1.0.0"
description = "Autonomous AI pipeline for cultural archiving of Igbo historical data."
readme = "README.md"
requires-python = "==3.13.*"
license = { text = "MIT" }
dependencies = [
    "google-adk",
    "python-dotenv",
    "python-telegram-bot",
    "huggingface-hub",
    "requests",
    "psycopg2-binary",
    "httpx",
    "pydantic",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents"]

[tool.uv]
package = true
